+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Yii Test Support Change Log
 
 
-## 1.1.1 under development
+## 1.2.0 February 23, 2021
 
 - Enh #24: Add `SimpleEventDispatcher::getEventClasses()` that return classes of dispatched events (vjik)